# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fulcrum Core API
  description: A comprehensive cloud infrastructure management system designed to orchestrate and monitor distributed cloud resources across multiple providers.
  version: 1.0.0
  contact:
    name: Fulcrum Project
    url: https://fulcrumproject.org
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers:
  - url: https://api.fulcrum.testudosrl.dev/api/v1
    description: Development server

# Default security for all operations
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: oauth2
      description: Role-based authentication using bearer tokens
      flows:
        implicit:
          authorizationUrl: https://api.fulcrum.testudosrl.dev/oauth/authorize
          scopes:
            admin: System administrator with unrestricted access
            participant: Participant administrator that can act as both provider and consumer
            agent: Agent role with access to specific operations
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
          example: "The field 'name' is required"
    properties.UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    JSONObject:
      type: object
      additionalProperties: true
      description: A generic properties.JSON object

    # Participant schemas
...
    ParticipantResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Test Participant"
        status:
          $ref: "#/components/schemas/ParticipantStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ParticipantStatus:
      type: string
      enum: [Enabled, Disabled]

    # Token schemas
    AuthRole:
      type: string
      enum: [admin, participant, agent]
      description: Access role for the token
...
    TokenResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "API Access Token"
        role:
          $ref: "#/components/schemas/AuthRole"
        expireAt:
          type: string
          format: date-time
        scopeId:
          $ref: "#/components/schemas/properties.UUID"
          description: "For participant role - the participant ID; for agent role - the agent ID"
        agentId:
          $ref: "#/components/schemas/properties.UUID"
          description: "For agent role tokens only - the agent ID"
        participantId:
          $ref: "#/components/schemas/properties.UUID"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        value:
          type: string
          description: "Plain token value. Only returned during token creation or regeneration"
          example: "eyJhbGciOiJIUzI1NiIsInR5c..."

    # Agent schemas
    AgentStatus:
      type: string
      enum: [New, Connected, Disconnected, Error, Disabled]
    AgentRequest:
      type: object
      required:
        - name
        - status
        - participantId
        - agentTypeId
      properties:
        name:
          type: string
          example: "aws-agent-01"
        status:
          $ref: "#/components/schemas/AgentStatus"
        properties:
          $ref: "#/components/schemas/JSONObject"
        participantId:
          $ref: "#/components/schemas/properties.UUID"
        agentTypeId:
          $ref: "#/components/schemas/properties.UUID"
    AgentResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "aws-agent-01"
        status:
          $ref: "#/components/schemas/AgentStatus"
        providerId:
          $ref: "#/components/schemas/properties.UUID"
        agentTypeId:
          $ref: "#/components/schemas/properties.UUID"
        tags:
          type: array
          items:
            type: string
          example: ["gpu", "high-memory", "ssd"]
        participant:
          $ref: "#/components/schemas/ParticipantResponse"
        agentType:
          $ref: "#/components/schemas/AgentTypeResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentCreateResponse:
      allOf:
        - $ref: "#/components/schemas/AgentResponse"
        - type: object
          properties:
            token:
              $ref: "#/components/schemas/TokenResponse"
              description: "The agent authentication token"

    # Agent Type schemas
...
    AgentTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Participant VM Agent"
        serviceTypes:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeResponse"
          description: "Service types that this agent type can provide"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service schemas
...
    ServiceTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "VM"
        propertySchema:
          $ref: "#/components/schemas/PropertySchema"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service Group schemas
...
    PageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        totalItems:
          type: integer
        totalPages:
          type: integer
        pageSize:
          type: integer
        page:
          type: integer
    PropertySchema:
      type: object
      description: properties.JSON schema defining validation rules for service properties
      additionalProperties:
        $ref: "#/components/schemas/PropertyDefinition"
    PropertyDefinition:
      type: object
      properties:
        type:
          type: string
          enum: [string, integer, number, boolean, object, array]
          description: The data type of the property
        label:
          type: string
          description: Human-readable label for the property
        required:
          type: boolean
          description: Whether the property is required
          default: false
        default:
          description: Default value for the property
        validators:
          type: array
          items:
            $ref: "#/components/schemas/ValidatorDefinition"
          description: Array of validation rules
        properties:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PropertyDefinition"
          description: Nested properties for object types
        items:
          $ref: "#/components/schemas/PropertyDefinition"
          description: Schema for array items
    ValidatorDefinition:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum:
            [
              minLength,
              maxLength,
              pattern,
              enum,
              min,
              max,
              minItems,
              maxItems,
              uniqueItems,
            ]
          description: Type of validator
        value:
          description: Value for the validator (type depends on validator type)
...
paths:
...
  /agents:
    get:
      summary: List agents
      tags:
        - Agents
      description: Retrieves a paginated list of agents
      x-auth-permissions:
        - role: admin
          permission: all agents
        - role: participant
          permission: agents belonging to its participant
        - role: agent
          permission: itself only
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: A paginated list of agents
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/AgentResponse"
    post:
      summary: Create an agent
      tags:
        - Agents
      description: Creates a new agent and returns the agent data with an authentication token
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: for its participant (when acting as provider)
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRequest"
      responses:
        "201":
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCreateResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    get:
      summary: Get an agent
      tags:
        - Agents
      description: Retrieves a specific agent by ID
      x-auth-permissions:
        - role: admin
          permission: all agents
        - role: participant
          permission: agents belonging to its participant
        - role: agent
          permission: itself only
      responses:
        "200":
          description: The agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update an agent
      tags:
        - Agents
      description: Updates an existing agent
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: agents belonging to its participant
        - role: agent
          permission: update its own status only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRequest"
      responses:
        "200":
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete an agent
      tags:
        - Agents
      description: Deletes an agent by ID
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: agents belonging to its participant
        - role: agent
          permission: not authorized
      responses:
        "204":
          description: Agent deleted successfully
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/me/status:
    put:
      summary: Update agent status
      tags:
        - Agents
      description: Updates the status of the authenticated agent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/AgentStatus"
      responses:
        "200":
          description: Agent status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/me:
    get:
      summary: Get current agent
      tags:
        - Agents
      description: Retrieves information about the currently authenticated agent
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  #
  # Agent Type endpoints
  #
...